fix: recent에 관련된 apis로직 server폴더로 이전
