--- apps/client/mutations/post/index.ts
@@ -1,3 +1,0 @@
-export { default as useUploadPost } from "./useUploadPost";
-export { default as useUpdatePost } from "./useUpdatePost";
-export { default as useDeletePost } from "./useDeletePost";
