--- apps/client/mutations/recent/index.ts
@@ -1,7 +1,0 @@
-export { default as useDeleteAllRecentHistory } from "./useDeleteAllRecentHistory";
-export { default as useDeleteRecentHistory } from "./useDeleteRecentHistory";
-export { default as useFetchDefaultMap } from "./useFetchDefaultMap";
-export { default as useFetchEatingHabitByFiltering } from "./useFetchEatingHabitByFiltering";
-export { default as useFetchRestaurantByCountry } from "./useFetchRestaurantByCountry";
-export { default as useFetchRestaurantNameByFiltering } from "./useFetchRestaurantNameByFiltering";
-export { default as useUploadRecentHistory } from "./useUploadRecentHistory";
