--- apps/client/mutations/recent/useUploadRecentHistory.ts
@@ -1,8 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-import { addRecentHistory } from "~/apis/recent";
-
-const useUploadRecentHistory = () => {
-  return useMutation(addRecentHistory);
-};
-
-export default useUploadRecentHistory;
