--- apps/client/queries/recent/keys.ts
@@ -1,5 +1,0 @@
-const Keys = {
-  recent: () => ["recent"]
-};
-
-export default Keys;
