--- apps/client/queries/recent/useFetchRecentHistory.ts
@@ -1,9 +1,0 @@
-import { getRecentHistory } from "~/apis/recent";
-import { useSuspendedQuery } from "~/hooks";
-import Keys from "./keys";
-
-const useFetchRecentHistory = () => {
-  return useSuspendedQuery(Keys.recent(), getRecentHistory);
-};
-
-export default useFetchRecentHistory;
