+++ apps/client/server/recent/keys.ts
@@ -1,0 +1,5 @@
+const Keys = {
+  recent: () => ["recent"]
+};
+
+export default Keys;
